requirements.txt
pbr>=0.6,!=0.7,<1.0
argparse
iso8601>=0.1.9
PrettyTable>=0.7,<0.8
python-keystoneclient>=0.7.0
six>=1.6.0
stevedore>=0.12

